% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_compare.R
\name{life_compare}
\alias{life_compare}
\title{Compare life expectancies}
\usage{
life_compare(
  file_path,
  file,
  status.include = 0,
  states,
  ref.var,
  ref.level,
  index.matrix,
  prop = TRUE,
  criterion = ">",
  status.names = NA
)
}
\arguments{
\item{file_path}{The file path for reading data.}

\item{file}{The file path for outputs.}

\item{status.include}{The status we aim to compare. It can be a number or a vector. Default is 0, which means we'll consider all status.}

\item{states}{The total number of transition states in our data.}

\item{ref.var}{A vector that contains all covariates used as reference variables when comparing each subgroup.}

\item{ref.level}{A vector that declares the reference value of each reference variable.}

\item{index.matrix}{A matrix that generated in \code{mlifeTable_plot()}. You don't need to specify it when using \code{mlifeTable_plot()}.}

\item{prop}{The indicator for proportion plots and tables. If TRUE, this function will output life expectancy proportion plots and tables in addition to orginal life expectancy plots. Default is TRUE.}

\item{criterion}{The criterion for comparison, which can be either ">" or "<". Default is ">".}

\item{status.names}{A vector used to specify names of each status except death.}
}
\value{
A \code{.csv} file.
}
\description{
A function for comparing life expectancies of each subgroup. By default, this function will calculate the proportion of samples that have a larger life expectancy or life expectancy proportion in your reference group than other groups.
}
\examples{
\dontrun{

#By setting the parameter 'compare' in mlifeTable_plot() to TRUE. We can directly use this function.
mlifeTable_plot(X=lifedata[,-1],status.include = 3,
      groupby = c("male","black","hispanic","other"), 
      cred = 0.92, 
      states = 3,
      file_path = ".",
      compare = TRUE,
      ref.var = c("black","hispanic","other"),
      ref.level = c(0,0,0))

}
}
