% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifedata.R
\docType{data}
\name{lifedata}
\alias{lifedata}
\title{Data of nominal brown shrimp prices by size}
\format{
A data frame with 1662 rows and 13 variables:
\describe{
  \item{trans}{Transition status was recorded}
  \item{age}{Age was recorded.}
  \item{male}{Sex. male=1, female=0}
  \item{black,other,hispanic}{Dummy variables for Race.}
  \item{edu}{Education level.}
  \item{region}{Region.}
  \item{immigrant}{Immigrant status.}
  \item{cohort}{Cohort.}
  \item{norcg}{norcg}
  \item{cogcg}{cogcg}
  \item{demcg}{demcg}
}
}
\usage{
lifedata
}
\description{
Shrimp are classified by size, 0-15 shrimp per pound, 15-20 shrimp per pound, etc. A smaller number per pound indicates larger shrimp. Nominal prices are total monthly value of brown shrimp andings within size class divided by total monthly landings within the size class.
}
\keyword{datasets}
