% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlifeTable.R
\name{mlifeTable}
\alias{mlifeTable}
\title{Multistate Life Table Method}
\usage{
mlifeTable(
  y,
  X,
  trans,
  states,
  file_path,
  groupby = NA,
  vars = colnames(X),
  status = 0,
  startages = 50,
  endages = 110,
  nums = 1000
)
}
\arguments{
\item{y}{A vector of transition results}

\item{X}{Covariates including age}

\item{trans}{The transition matrix generated by mlogit()}

\item{states}{The total number of transition states in our data.}

\item{file_path}{The file path for outputs.}

\item{groupby}{The covariates used for making subgroups.}

\item{vars}{The covariates considered in subgroup analysis.}

\item{status}{The status our simulated people are in.}

\item{startages}{start age of our life table.}

\item{endages}{end age of our life table.}

\item{nums}{number of life tables generated for each subgroup.}
}
\value{
Life tables for each subgroup.
}
\description{
A Multistate Life Table Method Based on Bayesian Approach
}
\examples{
data <- lifedata
y <- data[,1]
X <- data[,-1]
out <- mlogit(y, X ,samp=10, burn=10,verbose=10)

trans <- out
mlifeTable(y,X,trans =trans,
           groupby = c("male","norcg"),
           vars = "immigrant",
           states=9,
           nums = 5,
           file_path=".")
}
