% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlifeTable_plot.R
\name{mlifeTable_plot}
\alias{mlifeTable_plot}
\title{Plot life expectancies}
\usage{
mlifeTable_plot(
  status.include = 0,
  groupby,
  file_path,
  X,
  cred = 0.92,
  states,
  prop = TRUE,
  subgroup.names = NULL,
  status.names = NA,
  compare = FALSE,
  ...
)
}
\arguments{
\item{status.include}{The status we aim to calculate. It can be a number or a vector. Default is 0, which means we'll consider all status.}

\item{groupby}{A vector that contains all covariates used for making subgroups.}

\item{file_path}{The file path for outputs.}

\item{X}{Covariates including age. It should be a data.frame or a matrix.}

\item{cred}{credible level. e.g. if \code{cred = 0.92}, we will get the 92\% credible interval.}

\item{states}{The total number of transition states in our data.}

\item{prop}{The indicator for proportion plots and tables. If TRUE, this function will output life expectancy proportion plots and tables in addition to orginal life expectancy plots. Default is TRUE.}

\item{subgroup.names}{A vector that contains names of each subgroup. You can also specify them in the output files.}

\item{status.names}{A vector used to specify names of each status except death.}

\item{compare}{An indicator for generating comparsion results. If TRUE, this function will quote \code{life_compare()} and generate a table with all comparsion results based on the reference variables and reference levels you defined.}

\item{...}{Extra parameters for \code{life_compare()}. For details, please use \code{?life_compare()}.}
}
\value{
Plots and tables for posterior means and credible intervals of each subgroups
}
\description{
A Function for posterior means and credible intervals plots.
}
\examples{
\dontrun{

mlifeTable_plot(X=lifedata[,-1],status.include = 3,
      groupby = c("male","norcg","cogcg","demcg"), 
      cred = 0.92, 
      states = 9,
      file_path = ".")
      
}
}
