% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_compare.R
\name{life_compare}
\alias{life_compare}
\title{Compare life expectancies}
\usage{
life_compare(
  file_path,
  file = paste(file_path, "/mplotResults", sep = ""),
  state.include = 0,
  states,
  ref.var,
  ref.level,
  index.matrix,
  prop = TRUE,
  criterion = ">",
  state.names = NA
)
}
\arguments{
\item{file_path}{The file path for reading data. It can be inherited from \code{mlifetable_plot()}.}

\item{file}{The file path for outputs. Default is \code{paste(file_path,"/mplotResults",sep='')}.}

\item{state.include}{The status we aim to compare. It can be a number or a vector. Default is 0, which means we'll consider all states. It can be inherited from \code{mlifetable_plot()}.}

\item{states}{The total number of transition states in our data. It can be inherited from \code{mlifetable_plot()}.}

\item{ref.var}{A vector that contains all covariates used as reference variables when comparing each subgroup.}

\item{ref.level}{A vector that declares the reference value of each reference variable.}

\item{index.matrix}{A matrix that generated in \code{mlifeTable_plot()}. You don't need to specify it when using \code{mlifeTable_plot()}.}

\item{prop}{If TRUE, this function will output the comparision reulsts of life expectancy proportions in addition to orginal comparison results. Default is TRUE. It can be inherited from \code{mlifetable_plot()}.}

\item{criterion}{The criterion for comparison, which can be either ">" or "<". Default is ">".}

\item{state.names}{A vector used to specify names of each state except death. It can be inherited from \code{mlifetable_plot()}.}
}
\value{
A \code{.csv} file with comparison results.
}
\description{
A function for comparing life expectancies of each subgroup. By default, this function will calculate the proportion of samples that have a larger life expectancy or life expectancy proportion in your reference group than other groups.
}
\examples{
\dontrun{

#By setting the parameter 'compare' in mlifeTable_plot() to TRUE. We can directly use this function.
mlifeTable_plot(X=lifedata[,-1],state.include = 3,
      groupby = c("male","black","hispanic"), 
      cred = 0.92, 
      states = 3,
      file_path = ".",
      compare = TRUE,
      ref.var = c("black","hispanic"),
      ref.level = c(0,0))

}
}
