% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifedata.R
\docType{data}
\name{lifedata}
\alias{lifedata}
\title{Data for generating life tables.}
\format{
A data frame with 1662 rows and 13 variables:
\describe{
  \item{trans}{Transition status was recorded}
  \item{age}{Age was recorded.}
  \item{male}{Sex. male=1, female=0}
  \item{black,other,hispanic}{Dummy variables for Race.}
  \item{edu}{Education level.}
  \item{region}{Region.}
  \item{immigrant}{Immigrant status.}
  \item{cohort}{Cohort.}
  \item{norcg}{norcg}
  \item{cogcg}{cogcg}
  \item{demcg}{demcg}
}
}
\usage{
lifedata
}
\description{
Data for generating life tables.
}
\keyword{datasets}
