% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifedata.R
\docType{data}
\name{lifedata}
\alias{lifedata}
\title{Simplified Data for generating life tables.}
\format{
A data frame with 80146 rows and 16 variables:
\describe{
  \item{trans}{Transition status was recorded in the original data. In this data, we have 4 kinds of transtions in total.}
  \item{age}{Age for each subject.}
  \item{male}{Sex for each subject. male=1, female=0}
  \item{black,other,hispanic}{Dummy variables for Race.}
  \item{mar}{Marital status.}
  \item{educc,educg}{Dummy variables for education level.}
  \item{cohort}{Cohort. Birth cohort, which is birth year minus 1900. }
  \item{neb,mwb,wb}{Dummy variables for birth regions.}
  \item{nen,mwn,wn}{Dummy variables for residential regions.}
}
}
\source{
\url{https://hrsdata.isr.umich.edu/data-products/rand?_ga=2.225225498.1006069885.1653670364-1014684070.1647264850}
}
\usage{
lifedata
}
\description{
Data extracted and processed from The Health and Retirement Study (HRS).
}
\details{
Each subject has 3 status in the cohort: 1: health; 2: unhealthiness; 3: death. 
Thus we will have 6 kind of possible transitions: 1:health to health; 2:health to unhealthiness; 3: health to death; 4: unhealthiness to health; 5: unhealthiness to unhealthiness; 6: unhealthiness to death. To check the transition status for each subject, please use \code{lifedata[,1]}.

To use this package on your data, please make sure your data have a transition status vector.
}
\keyword{datasets}
