% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlifeTable.R
\name{mlifeTable}
\alias{mlifeTable}
\title{Multistate Life Table Method}
\usage{
mlifeTable(
  y,
  X,
  trans,
  states,
  file_path,
  groupby = NA,
  vars = NA,
  status = 0,
  startages = 50,
  endages = 110,
  age.gap = 2,
  nums = 1000,
  mlifeTable_plot = FALSE,
  status.names = NA,
  ...
)
}
\arguments{
\item{y}{A vector of transition results}

\item{X}{Covariates matrix. Note that X must include age as a convariate.}

\item{trans}{The transition matrix generated by mlogit()}

\item{states}{The total number of states in our data.}

\item{file_path}{The file path for outputs.}

\item{groupby}{The covariates used for making subgroups.}

\item{vars}{The covariates considered in subgroup analysis. For covariates that are not specified in \code{vars}, we will consider them have the same effect in each subgroup. Please make sure you have specified a variable otherwise the life tables would be the same.}

\item{status}{The status our simulated people are in. Default is 0, which means we will consider the general population.}

\item{startages}{start age of our life table. Default is 50.}

\item{endages}{end age of our life table. Default is 110.}

\item{age.gap}{The age gap when generating ages between each start age and end age. Default is 2.}

\item{nums}{number of life tables generated for each subgroup. Default is 1000.}

\item{mlifeTable_plot}{The indicator for generating tables and plots for posterior means and credible intervals. 
If TRUE, this function will create a new directory \code{mplotResults} under given \code{file_path} and output corresponding plots for posterior means and credible intervals.}

\item{status.names}{A vector used to specify names of each status except death. You can also specify them in the output files.}

\item{...}{Extra parameters for \code{mlifeTable_plot()}.For details, please use \code{?mlifeTable_plot()}}
}
\value{
Life tables for each subgroup.
}
\description{
A Multistate Life Table Method Based on Bayesian Approach. This function will return life tables for each subgroup. The values in the table represent the expectancy years for a subject to reach corresponding status. Note that the last column of each life table always indicates the expectancy years to death for each subject, while the order of other status in our life tables is determined by the order of status you specified when generating the transition status vector at the preprocessing step, which should be defined by yourself.
}
\examples{
\dontrun{
#The life tables generated in the example have 9 columns, which correspond to 9 status: 
#1: health; 2: hearts & stroke; 3: cancer; 
#4: hearts & cancer; 5: diabetes; 6: hearts & diabetes;
#7: diabetes & cancer; 8: Hearts, cancer & diabetes; 9: death.

data <- lifedata
y <- data[,1]
X <- data[,-1]
# this is a long running example
out <- mlogit(y, X ,samp=5000, burn=500,verbose=10)

trans <- out$outwstepwidth
mlifeTable(y,X,trans =trans,
           groupby = c("male","norcg"),
           vars = "immigrant",
           states=9,
           nums = 5,
           file_path=".")

#To name each status, try the status.names parameter.
mlifeTable(y,X,trans =trans,
           groupby = c("male","norcg"),
           vars = "immigrant",
           states=9,
           nums = 5,
           file_path=".",
           status.names= c("health","hearts & stroke","cancer",
                           "hearts & cancer", "diabetes", "hearts & diabetes",
                           "diabetes & cancer", "Hearts, cancer & diabetes"))
}
}
