% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlogit.R
\name{mlogit}
\alias{mlogit}
\title{Multistate Life Table Method}
\usage{
mlogit(
  y,
  X,
  m.0 = array(0, dim = c(ncol(X), ncol(y))),
  P.0 = array(0, dim = c(ncol(X), ncol(X), ncol(y))),
  samp = 1000,
  burn = 500,
  verbose = 1000
)
}
\arguments{
\item{y}{Dummy variables for multistates}

\item{X}{Covariates}

\item{m.0}{1}

\item{P.0}{1}

\item{samp}{Sampling times}

\item{burn}{'burn-in' times}

\item{verbose}{progress report}
}
\value{
An array for all parameter samples
}
\description{
A Multistate Life Table Method Based on Bayesian Approach
}
\examples{
mlogit(y, X, samp=3000, burn=0,verbose=10)
}
