% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlogit.R
\name{mlogit}
\alias{mlogit}
\title{Multistate Life Table Method}
\usage{
mlogit(
  y,
  X,
  file_path = NA,
  m.0 = array(0, dim = c(ncol(X), ncol(y))),
  P.0 = array(0, dim = c(ncol(X), ncol(X), ncol(y))),
  samp = 1000,
  burn = 500,
  verbose = 100,
  step.width = 5,
  trace.plot = FALSE
)
}
\arguments{
\item{y}{A vector of state transitions. For more details, please use \code{?lifedata}.}

\item{X}{Covariates matrix. Note that \code{X} must include age as a covariate.}

\item{file_path}{The file path for outputs. If a path is specified, the result will also be saved in the given file path. You can find two files in the specified file: \code{result.txt} and \code{resultwstep.txt}. The first one contains all samples generated after burn-in. The latter is sampled from the first one with a sampling interval.}

\item{m.0}{Initial matrix for MCMC methods.}

\item{P.0}{Initial matrix for MCMC methods.}

\item{samp}{Sampling times. We suggest a \code{samp} \eqn{\ge} 5000. To save time, we suggest you running multiple samplers in parallel.}

\item{burn}{'burn-in' times. To guarantee the convergence of our samples, we suggest a \code{burn} \eqn{\ge} 500.}

\item{verbose}{progress report. Default is 10, which means this function will report the current progress after every 10 samples.}

\item{step.width}{The sampling interval to reduce the autocorrelation. e.g. if \code{step.width = 5}, this function will pick one sample from every 5 samples and generate a new data set named \code{outwstepwidth.txt}. Default is 5.}

\item{trace.plot}{The indicator for generating trace plots. If TRUE, this function will create a new directory under given \code{file_path} and output corresponding trace plots using samples after burn-in.}
}
\value{
A list that constains two arrays:
\itemize{
  \item \strong{out}: An array that contains all samples generated.
  \item \strong{outwstepwidth}: An array generated by picking one sample from every \emph{step.width} samples in \strong{out}.
}
The number of columns in both arrays is determined by the number of covariates in X and the number of unique transition status in y. e.g. if we have 12 covariates in X and 36 unique transition status in y, our result will contain (12+1)*(36-1)= 455 columns in total.
}
\description{
A Multistate Life Table Method Based on Bayesian Approach.
}
\examples{
\dontrun{
data <- lifedata
y <- data[,1]
X <- data[,-1]

# this is a long running example. 
# To save time, you can run multiple samplers in parallel and combine the results.
# In practical application, please set burn at least 500 to guarantee the convergence.
out <- mlogit(y, X ,samp=250, burn=10,verbose=10)


}
}
