% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlogit.R
\name{mlogit}
\alias{mlogit}
\title{Multistate Life Table Method}
\usage{
mlogit(
  y,
  X,
  file_path = NA,
  m.0 = array(0, dim = c(ncol(X), ncol(y))),
  P.0 = array(0, dim = c(ncol(X), ncol(X), ncol(y))),
  samp = 1000,
  burn = 500,
  verbose = 1000
)
}
\arguments{
\item{y}{A vector of state transitions}

\item{X}{Covariates}

\item{file_path}{The file path for outputs.}

\item{m.0}{Initial matrix.}

\item{P.0}{Initial matrix.}

\item{samp}{Sampling times}

\item{burn}{'burn-in' times}

\item{verbose}{progress report}
}
\value{
An array for all parameter samples
}
\description{
A Multistate Life Table Method Based on Bayesian Approach
}
\examples{
data <- lifedata
y <- data[,1]
X <- data[,-1]
out <- mlogit(y, X ,samp=10, burn=10,verbose=10)
}
